dia6/deploy-render/requirements.txt
django==3.2
djangorestframework
mysqlclient
django-cors-headers
cloudinary
djangorestframework-simplejwt
drf-yasg
python-decouple
gunicorn
whitenoise